update component infoscreen
